+++ src/__tests__/Search/SearchComponent.test.tsx
@@ -1,7 +1,6 @@
 import { act } from 'react-dom/test-utils';
 
 import SearchComponent from '../../components/Search/SearchComponent';
-import SearchResultsList from '../../components/Search/SearchResultsList';
 import { Strings } from '../../resources/Strings';
 import { InputType } from '../../types/state';
 import { renderWithRouter } from '../utils/setupTests';
@@ -14,33 +13,10 @@ describe('Search View', () => {
       searchType: 'base' as InputType,
       mode: 'light' as 'light' | 'dark',
       view: 'search' as 'search' | 'compare-results',
+      isWarning: false,
       ...stringsBase,
     };
     renderWithRouter(<SearchComponent {...SearchPropsBase} />);
-
-    expect(document.body).toMatchSnapshot();
-    await act(async () => void jest.runOnlyPendingTimers());
-  });
-
-  it('displays selected revisions when user checks a revision', async () => {
-    const view = 'search' as 'search' | 'compare-results';
-    const mode = 'light' as 'light' | 'dark';
-    const searchType = 'base' as InputType;
-    const SearchPropsBase = {
-      searchType,
-      mode,
-      view,
-      ...stringsBase,
-    };
-
-    await act(async () => {
-      renderWithRouter(
-        <>
-          <SearchComponent {...SearchPropsBase} />,
-          <SearchResultsList view={view} mode={mode} searchType={searchType} />,
-        </>,
-      );
-    });
 
     expect(document.body).toMatchSnapshot();
     await act(async () => void jest.runOnlyPendingTimers());
