+++ src/__tests__/Search/SearchView.test.tsx
@@ -226,6 +226,7 @@ describe('Base Search', () => {
       searchType,
       mode: 'light' as 'light' | 'dark',
       view: 'search' as 'search' | 'compare-results',
+      isWarning: false,
       ...stringsBase,
     };
     renderComponent();
